I initialised the React app for the frontend and set up the structure for it under /client with essential react files and folders to provide scalability. src/components: contains reuseable UI components and src/services: handles API communication. I implemented src/services/flashcardService.js to manage API communication with the backend. It will contain the following functions: getFlashcards: Fetches all flashcards from the backend. addFlashcard: Adds a new flashcard with title and content. updateFlashcard: Updates the visibility of a flashcard by toggling its hidden status. deleteFlashcard: Deletes a specific flashcard using its ID.
For the components, i will utilise the following functions:
FlashcardList.js:
Displays a list of flashcards fetched from the backend.
Includes options to hide/unhide flashcards and delete them.

AddFlashcardForm.js:
Provides a form for adding new flashcards with fields for title and content.
Handles form submission and validation.

The react environment has been setup successfully and a check was carried out to check this.